Success test of request, saving it to database and getting the entity

+++ app/src/main/java/com/example/cookingrecipes/domain/Ingredient.kt
@@ -3,7 +3,7 @@ package com.example.cookingrecipes.domain
 data class Ingredient(
     val id: Int,
     val name: String,
-    val amount: Int,
+    val amount: Float,
     val measure: String
 )
